Major-project-main/requirements.txt
Flask==3.0.1
requests==2.32.3
scikit-learn==1.6.1
joblib==1.4.2
python-dotenv==1.0.1


